backend/requirements.txt
# MUSE Platform Production Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Mathematical and Scientific Computing
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.0
pandas==2.0.3

# Text Processing and NLP
syllapy==0.8.1
pronouncing==0.2.0
nltk==3.8.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1

# Caching and Session Management
redis==4.6.0
python-multipart==0.0.6

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Monitoring and Logging
prometheus-client==0.17.1
structlog==23.2.0

# HTTP Client
httpx==0.25.0
aiofiles==23.2.1

# Development and Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
black==23.9.1
isort==5.12.0
mypy==1.6.1

# Production Server
gunicorn==21.2.0